+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Handle multiple items in mainEntity as array defined in WebPage correctly (#25)
+
+### Deprecated
+- SchemaManager->setMainEntityOfWebPage() in favour of SchemaManager->addMainEntityOfWebPage() (#25)
+
 ## [1.4.0] - 2019-11-23
 
 ### Changed